+++ src/common/utils.c
@@ -244,12 +244,12 @@ bc_str_strip(char *str)
 
 
 char**
-bc_str_split(const char *str, char c, unsigned int max_pieces)
+bc_str_split(const char *str, char c, size_t max_pieces)
 {
     if (str == NULL)
         return NULL;
     char **rv = bc_malloc(sizeof(char*));
-    unsigned int i, start = 0, count = 0;
+    size_t i, start = 0, count = 0;
     for (i = 0; i < strlen(str) + 1; i++) {
         if (str[0] == '\0')
             break;
