+++ tests/common/check_utils.c
@@ -906,7 +906,7 @@ test_trie_size(void **state)
 }
 
 
-static unsigned int counter;
+static size_t counter;
 static char *expected_keys[] = {"chu", "copa", "bola", "bote", "bo", "b", "test", "testa"};
 static char *expected_datas[] = {"nda", "bu", "guda", "aba", "haha", "c", "asd", "lol"};
